images: draw inline images at their true fractional cell extent
An inline image's footprint was `div_ceil(pixels / cell)`, so a 402px sixel
on 21px cells (19.14 rows) was reported to the viewer as a full 20 cells.
The cursor is left on the image's last row, and the viewer's contain-fit then
scaled the image to fill that whole ceil'd row — so the prompt printed there
drew on top of the image ("passes through"). The real terminal keeps a 1:1
pixel mapping and leaves the fractional last row mostly empty for the text.

Emit the true fractional extent (pixels ÷ cell, un-rounded) as the wire display
size so the viewer draws the image at the same grid fraction the terminal did;
the partial last row stays empty for the prompt. The ceil is still tagged in
whole cells for the image's grid lifetime (scroll/erase/eviction). The viewer
already multiplies w/h as numbers, so no renderer change is needed.

Wire w/h change from u16 to f32 → SALT v6 (a v5 hub would drop image frames on
a value like 19.14 instead of failing loudly at register).

+++ src/model.rs
@@ -121,7 +121,7 @@ pub(crate) fn is_false(b: &bool) -> bool {
 /// in-frame base64 could never be cached and multiplied by every viewer).
 /// The bytes themselves live in [`Grid::image_data`] on the producing side
 /// and travel client→hub as blob messages, never inside frames.
-#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct ImagePlacement {
     /// Top-left cell of the image. May be negative when the image has partially
     /// scrolled off the top: the viewer clips it above the screen edge.
@@ -129,12 +129,15 @@ pub struct ImagePlacement {
     pub row: i16,
     #[serde(rename = "c")]
     pub col: u16,
-    /// Display size in cells, if the app specified one (else the browser uses the
-    /// image's natural pixel size).
+    /// Display size in cells — **fractional**. When the app specified a size
+    /// it is that exact integer; when derived from the image's pixels it is
+    /// `pixels ÷ cell-size` un-rounded, so the viewer draws the image at the
+    /// same grid fraction the terminal did (the partial last row stays partly
+    /// empty, instead of a ceil'd full row the cursor would then overlap).
     #[serde(rename = "w", default, skip_serializing_if = "Option::is_none")]
-    pub cols: Option<u16>,
+    pub cols: Option<f32>,
     #[serde(rename = "h", default, skip_serializing_if = "Option::is_none")]
-    pub rows: Option<u16>,
+    pub rows: Option<f32>,
     /// Content address of the image bytes ([`crate::proto::content_key`]).
     #[serde(rename = "k")]
     pub hash: String,
